Update Catalyst type

+++ src/main/scala/io/github/spark_redshift_community/spark/redshift/RedshiftJDBCWrapper.scala
@@ -477,7 +477,7 @@ private[redshift] class JDBCWrapper extends Serializable {
       case java.sql.Types.NUMERIC       => DecimalType(38, 18) // Spark 1.5.0 default
       // Redshift Real is represented in 4 bytes IEEE Float. https://docs.aws.amazon.com/redshift/latest/dg/r_Numeric_types201.html
       case java.sql.Types.REAL          => if (params.exists(_.legacyJdbcRealTypeMapping)) { DoubleType } else { FloatType }
-      case java.sql.Types.SMALLINT      => IntegerType
+      case java.sql.Types.SMALLINT      => ShortType
       case java.sql.Types.TINYINT       => IntegerType
       case _                            => null
       // scalastyle:on
